Reorganize and add more tests
